TIMOB-9141 BlackBerry: Fix merge problems
Fix some issues coming from merges
Also modify the population of the functionMap so order won't be an issue anymore.
- Use a vector for the map
- use enum instead of defines for the PROP ids, will give a bit more robustness,
  particularly protect against duplicates in prop names and in values
  Also allows keeping the prop in alphabetical order

Changelog:
- Use vector for the functionMap, added a static initializer
- Changed the type of the first param of setPropertyValue to size_t to prevent compiler warning about signed/unsigned comparison
- Removed left over setTop(float) declaration
- use enum for N_PROP ids
- use props everywhere in setTiMappingProperties
- updated app.js to use title for button text
- Commented a call that doesn't compile on R4
- Removed icon assets from bar-descriptor that don't exist anymore

Tests:
- Built tibb lib for all 3 targets
- Built tibbtest for all 3 targets
- Validated that app.js still works as before and labels do get displayed
- Tested everything with R4 tools

+++ NativeObject.cpp
@@ -17,7 +17,7 @@ NativeObject::~NativeObject()
 {
 }
 
-int NativeObject::setPropertyValue(int propertyNumber, TiObject* obj)
+int NativeObject::setPropertyValue(size_t propertyNumber, TiObject* obj)
 {
     return NATIVE_ERROR_NOTSUPPORTED;
 }
